+++ checkpoint/orbax/checkpoint/sharding_metadata.py
@@ -18,7 +18,8 @@ import abc
 import dataclasses
 import enum
 import json
-from typing import List, Tuple, Union
+import logging
+from typing import List, Optional, Tuple, Union
 import jax
 import numpy as np
 
@@ -50,7 +51,7 @@ def _convert_jax_partition_spec_to_partition_spec_elements(
       converted_element = None
     elif isinstance(element, str):
       converted_element = element
-    elif isinstance(element, tuple):
+    elif isinstance(element, (tuple, list)):
       converted_element = tuple(element)
     else:
       raise ValueError(f'Unsupported element type: {type(element)}')
@@ -158,7 +159,7 @@ class NamedShardingMetadata(ShardingMetadata):
 
   def __eq__(self, other):
     return (
-        self.shape == other.shape
+        np.array_equal(self.shape, other.shape)
         and self.axis_names == other.axis_names
         and self.partition_spec == other.partition_spec
     )
@@ -222,15 +223,15 @@ class PositionalShardingMetadata(ShardingMetadata):
   pass
 
 
-def from_jax_sharding(jax_sharding) -> ShardingMetadata:
+def from_jax_sharding(jax_sharding) -> Optional[ShardingMetadata]:
   """Converts `jax.sharding.Sharding` to `ShardingMetadata`."""
   if isinstance(jax_sharding, jax.sharding.NamedSharding):
     return NamedShardingMetadata.from_jax_sharding(jax_sharding)
   elif isinstance(jax_sharding, jax.sharding.SingleDeviceSharding):
     return SingleDeviceShardingMetadata.from_jax_sharding(jax_sharding)
   else:
-    raise NotImplementedError(
-        f'Conversion for {type(jax_sharding)} has not been implemented.'
+    logging.warning(
+        'Conversion for %s has not been implemented.', type(jax_sharding)
     )
 
 
@@ -251,3 +252,10 @@ def from_serialized_string(serialized_str) -> ShardingMetadata:
         f'Conversion for {deserialized_dict[_SHARDING_TYPE]} has not been'
         ' implemented.'
     )
+
+
+def get_sharding_or_none(serialized_string):
+  try:
+    return from_serialized_string(serialized_string.item()).to_jax_sharding()
+  except ValueError as e:
+    logging.error(e)