+++ checkpoint/orbax/checkpoint/value_metadata.py
@@ -20,6 +20,7 @@ from typing import Optional
 from etils import epath
 import jax
 from jax import numpy as jnp
+from orbax.checkpoint import sharding_metadata
 
 
 @dataclasses.dataclass
@@ -47,15 +48,15 @@ class ArrayMetadata(Metadata):
   shape:
     Tuple of integers describing the array shape.
   sharding:
-    jax.sharding.Sharding to indicate how the array is sharded. In most of the
-    cases, it's NamedSharding.
-    May be None if the array is not sharded.
+    ShardingMetadata to indicate how the array is sharded. ShardingMetadata is
+    an orbax representation of `jax.sharding.Sharding` which stores the same
+    properties but not require accessing real devices.
   dtype:
     Dtype of array elements.
   """
 
   shape: tuple[int, ...]
-  sharding: Optional[jax.sharding.Sharding]
+  sharding: Optional[sharding_metadata.ShardingMetadata]
   dtype: Optional[jnp.dtype]
 
   def __eq__(self, other: 'Metadata') -> bool:
@@ -91,7 +92,7 @@ class ScalarMetadata(ArrayMetadata):
   """
 
   shape: tuple[int, ...] = tuple([])
-  sharding: Optional[jax.sharding.Sharding] = None
+  sharding: Optional[sharding_metadata.ShardingMetadata] = None
   dtype: Optional[jnp.dtype] = None
 
   def __eq__(self, other: 'Metadata') -> bool: